break out auto fields into two tests
