+++ tests/test_salesforce_automatic_fields_rest.py
@@ -77,19 +77,6 @@ class SalesforceAutomaticFields(SalesforceBaseTest):
                     self.assertSetEqual(expected_keys, actual_keys)
 
 
-class SalesforceAutomaticFieldsBulk(SalesforceAutomaticFields):
-    """Test that with no fields selected for a stream automatic fields are still replicated"""
-
-    salesforce_api = 'BULK'
-
-    @staticmethod
-    def name():
-        return "tt_salesforce_auto_bulk"
-
-    def test_run(self):
-        self.automatic_fields_test()
-
-
 class SalesforceAutomaticFieldsRest(SalesforceAutomaticFields):
     """Test that with no fields selected for a stream automatic fields are still replicated"""
